Smooth Spinners: Makes the loading "spinner" animation smoother in no…
…n-holo apps

The loading "spinner" animation is quite jerky by default when the
progress xml's from
drawable are called in applications that don't use the holo progress
animation. The holo
"spinner" animation is already smooth by default. This makes every other
"spinner"
animation smooth.

We've been changing this value in all ROM's we release, both AOSP and CM
based ones
as well as stock ROM's, since it gives a more polished look to the
entire system. In the past
3 years we've been using this small tweak we have not noticed anything
that would
adversely affect battery life or system operations as a whole.

Change-Id: I4c333ecaee72844fb4d06a3cc6925b17e2208433
Signed-off-by: Sar Castillo <sar.castillo@gmail.com>